# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend service for managing catalog and orders.",
        "title": "BooksOnline API",
        "contact": {},
        "version": "1.0"
    },
    "host": "api-bookstore.mydomain.com",
    "basePath": "/",
    "paths": {
        "/orders": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Returns all orders (Admin only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "List all orders",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/orders.Order"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates an order and generates digital keys for licenses/vouchers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Place a new order",
                "parameters": [
                    {
                        "description": "Order request",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/adapters.CreateOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/orders.Order"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/orders/{id}": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Returns order details (Admin only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Get order by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/orders.Order"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/products": {
            "get": {
                "description": "Returns the entire product catalog",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "catalog"
                ],
                "summary": "List products",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/catalog.Product"
                            }
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "description": "Returns a single product detail by UUID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "catalog"
                ],
                "summary": "Get product by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/catalog.Product"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "adapters.CreateOrderRequest": {
            "type": "object",
            "properties": {
                "customer_email": {
                    "type": "string"
                },
                "items": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                }
            }
        },
        "catalog.Product": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "isActive": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "priceInCents": {
                    "type": "integer",
                    "format": "int64"
                },
                "sku": {
                    "type": "string"
                },
                "stock": {
                    "type": "integer"
                },
                "type": {
                    "$ref": "#/definitions/catalog.ProductType"
                }
            }
        },
        "catalog.ProductType": {
            "type": "string",
            "enum": [
                "BOOK",
                "LICENSE",
                "VOUCHER"
            ],
            "x-enum-varnames": [
                "TypeBook",
                "TypeLicense",
                "TypeVoucher"
            ]
        },
        "orders.Order": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "customerEmail": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/orders.OrderItem"
                    }
                },
                "status": {
                    "$ref": "#/definitions/orders.OrderStatus"
                },
                "totalPrice": {
                    "type": "integer",
                    "format": "int64"
                }
            }
        },
        "orders.OrderItem": {
            "type": "object",
            "properties": {
                "generatedAsset": {
                    "description": "For License or Voucher this holds the generated key/code",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "orderID": {
                    "type": "string"
                },
                "price": {
                    "type": "integer",
                    "format": "int64"
                },
                "productID": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "orders.OrderStatus": {
            "type": "string",
            "enum": [
                "PENDING",
                "COMPLETED",
                "FAILED"
            ],
            "x-enum-varnames": [
                "StatusPending",
                "StatusCompleted",
                "StatusFailed"
            ]
        }
    }
}